Modify tonic build so it builds to src/
* Allows IntelliJ to provide code completion
* Generated files are ignored
