+++ factory_functional_units/build.rs
@@ -1,4 +1,7 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("proto/functional_units.proto")?;
+    tonic_build::configure()
+        .build_client(false)
+        .out_dir("src/server")
+        .compile(&["proto/functional_units.proto"], &["proto"])?;
     Ok(())
 }
